LayoutTests/fast/profiler/simple-no-level-change-expected.txt

CONSOLE MESSAGE: line 59: A simple profile test where no scope chagnes
CONSOLE MESSAGE: line 75: Thread_1  0
CONSOLE MESSAGE: line 75:  functionWichStartsAndStopsTheProfiler simple-no-level-change.html 14
CONSOLE MESSAGE: line 75:   getElementById  0
CONSOLE MESSAGE: line 75:  (idle)  0
This page's JavaScript starts and stops profiling from the same scope. 

To run this test manually, load it in the browser then load the WebInspector and look at the profile. There should be no function calls between console.profile() and console.profileEnd().
